Fixed bugs with framrate dependency, added disblocks and keys

+++ Assets/Scripts/ScreenShake/IScreenShakeService.cs
@@ -6,7 +6,7 @@ public interface IScreenShakeService
 {
     void GetCamera();
     void StartScreenShake(float time, float amount);
-    void StartTransition(int fadeTime, bool fadeIn);
-    void StartScreenFlash(int time, float amount);
+    void StartTransition(float fadeTime, bool fadeIn);
+    void StartScreenFlash(float time, float amount);
     void StartSwipe(bool fadeIn);
 }
